[Bronze V] Title: A/B, Time: 44 ms, Memory: 31256 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.py
@@ -1,3 +1,3 @@
 import sys
-a,b=map(float,sys.stdin.readline().split())
+a,b=map(int,sys.stdin.readline().split())
 print(a/b)
